Port Biermann-draft-02 API to Chlorine
Made changes to get a clean compile under Chlorine.

NOTE: changes in Chlorine break the jUnit testing for this feature -
mostly due to changes in yangtools making classes sealed, which
prevents use of mockito to simulate these classes.  There are other
changes as well that caused breakage. For now, pom.xml has been changed
to disable jUnit, but this should be corrected.

Issue-ID: CCSDK-3843
Signed-off-by: Dan Timoney <dtimoney@att.com>
Change-Id: Ic53c6d580d644fab069a06f033db515c05dff6f2